+++ lib/src/preloaded_image/raw_preloaded_image.dart
@@ -1,4 +1,5 @@
 import "package:flutter/material.dart";
+import "package:flutter/scheduler.dart";
 
 import "preloaded_image.dart";
 
@@ -15,7 +16,7 @@ typedef RawPreloadedImageBuilder = Widget Function(FittedSizes? sizes, BoxConstr
 class RawPreloadedImage extends StatefulWidget {
   final ImageProvider image;
 
-  final ImageConfiguration configuration;
+  final ImageConfiguration? configuration;
 
   /// called with all parameter set to null if the image is not loaded
   /// and only with values if the image was loaded
@@ -39,13 +40,12 @@ class RawPreloadedImage extends StatefulWidget {
   const RawPreloadedImage({
     super.key,
     required this.image,
-    ImageConfiguration? configuration,
+    this.configuration,
     required this.builder,
     this.errorBuilder,
     this.chunkBuilder,
     BoxFit? boxFit,
-  })  : configuration = configuration ?? ImageConfiguration.empty,
-        boxFit = boxFit ?? BoxFit.contain;
+  }) : boxFit = boxFit ?? BoxFit.contain;
 
   @override
   State<RawPreloadedImage> createState() => _RawPreloadedImageState();
@@ -53,15 +53,17 @@ class RawPreloadedImage extends StatefulWidget {
 
 class _RawPreloadedImageState extends State<RawPreloadedImage> {
   ImageStream? _stream;
+  ImageInfo? _imageInfo;
 
   Size? _size;
 
-  late Widget _child;
-
+  late DisposableBuildContext<State<RawPreloadedImage>> _scrollAwareContext;
   _ImageState _imageState = _ImageState.LOADING;
 
+  late Widget _child;
+
   late final ImageStreamListener _listener = ImageStreamListener(
-    _onBuildChild,
+    _onBuildImage,
     onError: widget.errorBuilder != null ? _onBuildError : null,
     onChunk: widget.chunkBuilder != null ? _onBuildChunk : null,
   );
@@ -83,8 +85,9 @@ class _RawPreloadedImageState extends State<RawPreloadedImage> {
   void initState() {
     super.initState();
 
+    _scrollAwareContext = DisposableBuildContext<State<RawPreloadedImage>>(this);
+
     _buildChild();
-    _requestImage();
   }
 
   @override
@@ -92,7 +95,7 @@ class _RawPreloadedImageState extends State<RawPreloadedImage> {
     super.didUpdateWidget(oldWidget);
 
     if (oldWidget.image != widget.image || oldWidget.configuration != widget.configuration) {
-      _requestImage();
+      _resolveImage();
     }
 
     if (oldWidget.builder != widget.builder || oldWidget.boxFit != widget.boxFit) {
@@ -106,8 +109,52 @@ class _RawPreloadedImageState extends State<RawPreloadedImage> {
     }
 
     if (oldWidget.errorBuilder != widget.errorBuilder || oldWidget.chunkBuilder != widget.chunkBuilder) {
-      _requestImage();
+      _resolveImage();
     }
+  }
+
+  @override
+  void didChangeDependencies() {
+    _resolveImage();
+
+    super.didChangeDependencies();
+  }
+
+  void _resolveImage() {
+    ScrollAwareImageProvider provider = ScrollAwareImageProvider<Object>(
+      context: _scrollAwareContext,
+      imageProvider: widget.image,
+    );
+    ImageStream imageStream = provider.resolve(widget.configuration ?? createLocalImageConfiguration(context));
+
+    _updateSourceStream(imageStream);
+  }
+
+  void _updateSourceStream(ImageStream stream) {
+    if (_stream?.key == stream.key) return;
+
+    if (_stream != null) {
+      _stream!.removeListener(_listener);
+      _stream!.removeListener(_stateListener);
+    }
+
+    stream.addListener(_stateListener);
+    stream.addListener(_listener);
+
+    _stream = stream;
+  }
+
+  @override
+  void reassemble() {
+    // in case the image cache was flushed
+    _resolveImage();
+    super.reassemble();
+  }
+
+  void _replaceImage({required ImageInfo? info}) {
+    ImageInfo? o = _imageInfo;
+    SchedulerBinding.instance.addPostFrameCallback((_) => o?.dispose(), debugLabel: "RawPreloadedImage.disposeOldInfo");
+    _imageInfo = info;
   }
 
   @override
@@ -115,16 +162,8 @@ class _RawPreloadedImageState extends State<RawPreloadedImage> {
     return _child;
   }
 
-  void _requestImage() {
-    ImageStream stream = widget.image.resolve(widget.configuration);
-
-    stream.removeListener(_listener);
-    stream.removeListener(_stateListener);
-
-    stream.addListener(_stateListener);
-    stream.addListener(_listener);
-
-    _stream = stream;
+  void _buildChild() {
+    _child = _size == null ? widget.builder(null, null) : _buildLoadedImageChild();
   }
 
   /// the loaded image as a widget.
@@ -140,13 +179,10 @@ class _RawPreloadedImageState extends State<RawPreloadedImage> {
     );
   }
 
-  void _buildChild() {
-    _child = _size == null ? widget.builder(null, null) : _buildLoadedImageChild();
-  }
-
-  void _onBuildChild(ImageInfo image, bool synchronousCall) {
+  void _onBuildImage(ImageInfo image, bool synchronousCall) {
     if (context.mounted) {
       setState(() {
+        _replaceImage(info: image);
         _size = Size(image.image.width.toDouble(), image.image.height.toDouble());
         _buildChild();
       });
@@ -175,133 +211,8 @@ class _RawPreloadedImageState extends State<RawPreloadedImage> {
   void dispose() {
     _stream?.removeListener(_stateListener);
     _stream?.removeListener(_listener);
+    _replaceImage(info: null);
+    _scrollAwareContext.dispose();
     super.dispose();
   }
 }
-
-/*
-import "package:flutter/material.dart";
-
-import "raw_preloaded_image.dart";
-
-/// Preloads the image file to size the layout according to the image and available space.
-///
-/// This allows an ink animation to be exactly on the image and not spreading over it.
-class PreloadedImage extends StatelessWidget {
-  /// builder used to allow implementations of animations
-  ///
-  /// if [image] is null, the image is not loaded yet.
-  /// if [image] is not null, it represents the full widget to display the image with all defined settings.
-  final Widget Function(Widget? image) builder;
-
-  /// [DecorationImage.image] is the resolved image.
-  /// after the resolving, [image] is placed in [decoration].
-  /// The separation of the image is done for convenience.
-  final DecorationImage image;
-
-  final ImageConfiguration configuration;
-  final BorderRadius? borderRadius;
-
-  /// [BoxDecoration.borderRadius] and [BoxDecoration.image] are ignored
-  final BoxDecoration decoration;
-
-  /// The fit for the ink widget in his parent.
-  ///
-  /// defaults to [BoxFit.contain] which allows to size the ink animation to its [decoration]
-  ///
-  /// if set to [BoxFit.fill], it will expand the ink animation area but not the [decoration]
-  /// configure [DecorationImage.fit] to directly control how the image is sized inside its ink container.
-  final BoxFit boxFit;
-
-  /// only effects [child]
-  final EdgeInsets? padding;
-
-  /// positioned above the image
-  final Widget? child;
-
-  final void Function()? onPressed;
-
-  /// allows the full control of the image widget
-  /// all parameter are only effective on the loaded image.
-  const PreloadedImage.builder({
-    super.key,
-    required this.builder,
-    required this.image,
-    ImageConfiguration? configuration,
-    this.borderRadius,
-    BoxDecoration? decoration,
-    BoxFit? boxFit,
-    this.onPressed,
-    this.padding,
-    this.child,
-  })  : configuration = configuration ?? ImageConfiguration.empty,
-        decoration = decoration ?? const BoxDecoration(),
-        boxFit = boxFit ?? BoxFit.contain;
-
-  /// replaces the image with a [loading] widget until it is fully loaded
-  PreloadedImage({
-    Key? key,
-    required DecorationImage image,
-    ImageConfiguration? configuration,
-    BorderRadius? borderRadius,
-    BoxDecoration? decoration,
-    BoxFit? boxFit,
-    Widget? loading,
-    void Function()? onPressed,
-    EdgeInsets? padding,
-    Widget? child,
-  }) : this.builder(
-          key: key,
-          image: image,
-          configuration: configuration,
-          builder: (image) => _loadingBuilder(image, loading ?? _loading),
-          borderRadius: borderRadius,
-          decoration: decoration,
-          boxFit: boxFit,
-          onPressed: onPressed,
-          padding: padding,
-          child: child,
-        );
-
-  static const _loading = Center(child: CircularProgressIndicator());
-
-  static Widget _loadingBuilder(Widget? image, Widget loading) {
-    return image ?? loading;
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return RawPreloadedImage(
-      image: image.image,
-      builder: (sizes, constraints) {
-        if (sizes == null) return builder(null);
-
-        print("BUILD");
-        return builder(
-          InkWell(
-            borderRadius: borderRadius,
-            onTap: onPressed,
-            child: Ink(
-              width: sizes.destination.width,
-              height: sizes.destination.height,
-              padding: padding,
-              decoration: BoxDecoration(
-                color: decoration.color,
-                gradient: decoration.gradient,
-                shape: decoration.shape,
-                border: decoration.border,
-                boxShadow: decoration.boxShadow,
-                backgroundBlendMode: decoration.backgroundBlendMode,
-                borderRadius: borderRadius,
-                image: image,
-              ),
-              child: child,
-            ),
-          ),
-        );
-      },
-    );
-  }
-}
-
- */